requirements.txt
google-adk
litellm
python-dotenv